CLOUD-539 Deleted unused email service. Refactored the used one
